Move waiting for configmap to conftest
